+++ Mac-Old/Mac-Old/Views/JVMarkedScroller.m
@@ -1,3 +1,4 @@
+#include <tgmath.h>
 #import "JVMarkedScroller.h"
 
 @interface JVMark : NSObject
@@ -43,12 +44,12 @@
 	[super drawRect:rect];
 
 	NSAffineTransform *transform = [NSAffineTransform transform];
-	float width = [[self class] scrollerWidthForControlSize:[self controlSize] scrollerStyle:NSScrollerStyleOverlay];
+	CGFloat width = [[self class] scrollerWidthForControlSize:[self controlSize] scrollerStyle:NSScrollerStyleOverlay];
 
-	float scale = [self scaleToContentView];
+	CGFloat scale = [self scaleToContentView];
 	[transform scaleXBy:( sFlags.isHoriz ? scale : 1. ) yBy:( sFlags.isHoriz ? 1. : scale )];
 
-	float offset = [self rectForPart:NSScrollerKnobSlot].origin.y;
+	CGFloat offset = [self rectForPart:NSScrollerKnobSlot].origin.y;
 	[transform translateXBy:( sFlags.isHoriz ? offset / scale : 0. ) yBy:( sFlags.isHoriz ? 0. : offset / scale )];
 
 	NSBezierPath *shades = [NSBezierPath bezierPath];
@@ -113,8 +114,8 @@
 
 		NSPoint point = NSMakePoint( ( sFlags.isHoriz ? value : 0. ), ( sFlags.isHoriz ? 0. : value ) );
 		point = [transform transformPoint:point];
-		point.x = ( sFlags.isHoriz ? roundf( point.x ) + 0.5 : point.x );
-		point.y = ( sFlags.isHoriz ? point.y : roundf( point.y ) + 0.5 );
+		point.x = ( sFlags.isHoriz ? round( point.x ) + 0.5 : point.x );
+		point.y = ( sFlags.isHoriz ? point.y : round( point.y ) + 0.5 );
 
 		if( ! NSPointInRect( point, knobRect ) ) {
 			if( mark.color ) {
@@ -243,7 +244,7 @@
 	NSEvent *event = [[NSApplication sharedApplication] currentEvent];
 	NSPoint where = [self convertPoint:[event locationInWindow] fromView:nil];
 	NSRect slotRect = [self rectForPart:NSScrollerKnobSlot];
-	float scale = [self scaleToContentView];
+	CGFloat scale = [self scaleToContentView];
 	[self removeMarksLessThan:( ( sFlags.isHoriz ? where.x - NSMinX( slotRect ) : where.y - NSMinY( slotRect ) ) / scale )];
 }
 
@@ -251,7 +252,7 @@
 	NSEvent *event = [[NSApplication sharedApplication] currentEvent];
 	NSPoint where = [self convertPoint:[event locationInWindow] fromView:nil];
 	NSRect slotRect = [self rectForPart:NSScrollerKnobSlot];
-	float scale = [self scaleToContentView];
+	CGFloat scale = [self scaleToContentView];
 	[self removeMarksGreaterThan:( ( sFlags.isHoriz ? where.x - NSMinX( slotRect ) : where.y - NSMinY( slotRect ) ) / scale )];
 }
 
@@ -494,7 +495,7 @@
 }
 
 - (CGFloat) shiftAmountToCenterAlign {
-	float scale = [self scaleToContentView];
+	CGFloat scale = [self scaleToContentView];
 	if( sFlags.isHoriz ) return ( ( NSWidth( [self rectForPart:NSScrollerKnobSlot] ) * [self knobProportion] ) / 2. ) / scale;
 	else return ( ( NSHeight( [self rectForPart:NSScrollerKnobSlot] ) * [self knobProportion] ) / 2. ) / scale;
 }
